requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
Django==4.2.2
django-cors-headers==4.1.0
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.0
fonttools==4.39.4
idna==3.4
iniconfig==2.0.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.0.0
pydataset==0.2.0
PyJWT==2.7.0
pyparsing==3.0.9
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
stripe==5.5.0
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.23.0
